+++ firmware/src/worker.rs
@@ -276,7 +276,16 @@ impl<'a> Worker<'a> {
         kv_total: u32,
     ) -> Self {
         Self {
-            ctap: AppletHandler::new(fs, rng, presence, serial_id, serial_hash, otp_key, devk),
+            ctap: AppletHandler::new(
+                fs,
+                rng,
+                presence,
+                presence,
+                serial_id,
+                serial_hash,
+                otp_key,
+                devk,
+            ),
             ccid: CcidApplets::new(
                 fs,
                 rng,